changed review publish order and moved sku
