+++ Localize/lang/strings.json
@@ -27,6 +27,7 @@
   "/I/U8L": "Body",
   "/IVuGP": "Discard",
   "/KRvvg": "No values match your search.",
+  "/LO3Ia": "Testing",
   "/NtebP": "(UTC+05:00) Islamabad, Karachi",
   "/QqADs": "Returns the elements in the array starting at index Count",
   "/RS9F7": "(UTC-06:00) Saskatchewan",
@@ -1249,6 +1250,7 @@
   "_/I/U8L.comment": "Label text for request or response body",
   "_/IVuGP.comment": "Button text for discard the dialog",
   "_/KRvvg.comment": "Label for when no values match search value.",
+  "_/LO3Ia.comment": "The text for the testing environment",
   "_/NtebP.comment": "Time zone value ",
   "_/QqADs.comment": "Label for description of custom skip Function",
   "_/RS9F7.comment": "Time zone value ",
